pyproject.toml
[tool.poetry]
name = "myair"
version = "0.0.1"
description = "Pet project, which is a clone of the airline website"
authors = ["qq2"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
django = "^4.2.6"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
requests = "^2.31.0"
amadeus = "^9.0.0"
django-split-settings = "^1.3.1"
django-debug-toolbar = "^4.4.2"
pre-commit = "^3.7.1"
drf-spectacular = "^0.27.2"
mock = "^5.1.0"
kafka-python = "^2.0.2"
python-telegram-bot = "^21.3"
sqlalchemy = {extras = ["async"], version = "^2.0.31"}
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
